Refactor utilities into CLI utilities folder
