+++ clients/typescript/src/cli/util/index.ts
@@ -1,1 +1,4 @@
 export * from './io'
+export * from './parse'
+export * from './string'
+export * from './paths'
