+++ clients/typescript/src/cli/util/paths.ts
@@ -1,0 +1,6 @@
+import path from 'path'
+
+/**
+ * Path where the user ran `npx electric`
+ */
+export const appRoot = path.resolve()
